// src/services/types/platform/in_app_feedback.ts
export type InAppFeedbackApiResponse = {
  _id: string
  __v: number
  in_app_feedback: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
    }
    section2: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
    }
    section3_tab1: {
      image1: string
      title1_en: string
      title1_fa: string
      video1: string
    }
    section3_tab2: {
      image1: string
      title1_en: string
      title1_fa: string
      video1: string
    }
    section4: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      image1: string
    }
    section5: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
      image1: string
    }
    section6: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      image1: string
    }
    section7: {
      image1: string
      image2: string
      image3: string
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      title3_en: string
      title3_fa: string
      s_title1_en: string
      s_title1_fa: string
      s_title2_en: string
      s_title2_fa: string
      s_title3_en: string
      s_title3_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
    }
    section8: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
}